Updated my Datalogger code
Put the Power Measurement stuff in a separate file that we currently aren't using.
Cut down on the number of Arduino Functions we're using.

Using these Arduino Functions:
    3 x pinMode(...) to set the output pins.
    2 x Delay(...) for my LED Blink function.
    Serial.* for debugging (Currently disabled by a #define)
    SD.h for all of my SD card needs

I'm getting tired of the Arduino stuff, so it may go away soon.

Trying to get the ADC interrupt to work right, but I think the compilers optimization is eating some of my while loops.

--Arthur

git-svn-id: http://cd-net.net/CPE496/trunk@112 05b6d0e0-c345-4a8e-a5b7-579c8bd0b646