Code Update
I've included everything I use to program my board except the Arduino compiler.

It's looking like averaging is working instead of RMS for some unknown reason right now, but I need to get a variable voltage source/current sync and check this.

In other news the transformer seems to be putting out more voltage than it should.  It's supposed to be a 23:1 Ratio, but I was measuring 6.4V on the secondaries and my Kill-A-Watt was saying 123V on the primary.

TODO:
Get a variable voltage source/current sync and test the linearity of my measurements.
If they're linear, then I can account for any weird offsets, ratios.

--Arthur

git-svn-id: http://cd-net.net/CPE496/trunk@118 05b6d0e0-c345-4a8e-a5b7-579c8bd0b646