Modified Price List
This is the final price list for everything we need to make 3 boards.
It does not include the cost of 10AWG wire. (DigiKey wants us to buy to much)
Also included is a rubric for delivery time vs number of boards.
This will let us choose the best option for PCB manufacturing.

--Arthur

git-svn-id: http://cd-net.net/CPE496/trunk@131 05b6d0e0-c345-4a8e-a5b7-579c8bd0b646